+++ src/rules-browserslist.ts
@@ -5,7 +5,7 @@ import { openUri } from './utilities';
 import { exState } from './tree-provider';
 import { ignore } from './ignore';
 import { browsersList, exists } from './analyzer';
-import { existsSync, readFileSync, rmSync, writeFileSync } from 'fs';
+import { existsSync, readFileSync, writeFileSync } from 'fs';
 import { window } from 'vscode';
 import { join } from 'path';
 import { getPackageJSONFilename } from './monorepo';
@@ -37,17 +37,7 @@ export function checkBrowsersList(project: Project) {
       filename = join(folder, name);
     }
     if (exists('@angular/core')) {
-      if (existsSync(filename)) {
-        // Migrate to package.json
-        const title = 'Reduce Config Files';
-        const message = `${name} can be moved into your package.json`;
-        project.add(
-          new Tip(title, message, TipType.Idea)
-            .setQueuedAction(moveFile, project, name, filename, title, message)
-            .canIgnore(),
-        );
-        return;
-      } else {
+      if (!existsSync(filename)) {
         const title = 'Set Browser Support';
         const message = `Some older devices will not be supported. Updating your package.json to include browserslist will fix this.`;
         project.add(
@@ -98,45 +88,6 @@ async function setBrowsersList(queueFunction: QueueFunction, project: Project, t
     fixPackageJson(project, defaultValues());
   } catch (err) {
     window.showErrorMessage(`Failed to fix browserslist: ${err}`);
-  }
-}
-
-async function moveFile(
-  queueFunction: QueueFunction,
-  project: Project,
-  name: string,
-  filename: string,
-  title: string,
-  message,
-) {
-  const choice = await window.showInformationMessage(
-    `The file ${name} can be moved into package.json to reduce the number of config files in your project. Would you like to do this?`,
-    'Yes, Apply Changes',
-    'Ignore',
-  );
-  if (!choice) {
-    return;
-  }
-
-  try {
-    if (choice == 'Ignore') {
-      ignore(new Tip(title, message), exState.context);
-      return;
-    }
-    queueFunction();
-    const txt = readFileSync(filename, 'utf8').split(/\r?\n/);
-    const lines = txt.map((line) => line.trim());
-    const list = [];
-    for (const line of lines) {
-      if (line && !line.startsWith('#')) {
-        const arg = line.split('#')[0];
-        list.push(arg);
-      }
-    }
-    fixPackageJson(project, list);
-    rmSync(filename);
-  } catch (err) {
-    window.showErrorMessage(`Failed to fix ${name}: ${err}`);
   }
 }