+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### Version 2.1.6
+
+- Remove "Reduce Config Files" browserslist tip
+- Fix import of @webnativellc/simple-plist to properly access readFileSync and writeFileSync
+
 ### Version 2.1.5
 
 - Fix for "Cannot read properties of undefined (reading 'replace') at finishCommand"